\name{default_parameters}
\alias{default_parameters}
\title{Default ThermoPheno crop parameters}
\usage{
default_parameters(crop_type = c("summer", "winter"))
}
\arguments{
  \item{crop_type}{Crop type: summer or winter.}
}
\value{A named list of default crop parameters.}
\description{Default ThermoPheno crop parameters}
